changed bookings to academic-weeks
