+++ backend/treeckle/nusmods/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from . import views 
+
+urlpatterns = [
+    path("academic-weeks/", views.get_academic_weeks, name="academic-weeks"),
+]
